requirements.txt
# Archivo: requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Login==0.6.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.2.3